fix: parent class for duplicate code
