+++ src/core/light_curve/infrastructure/detection_repository.py
@@ -1,6 +1,5 @@
 from db_plugins.db.sql.connection import SQLConnection
 from db_plugins.db.mongo.connection import MongoConnection
-from db_plugins.db.sql import models as psql_models
 from db_plugins.db.mongo import models as mongo_models
 from returns.result import Success, Failure
 from returns.pipeline import is_successful
@@ -9,38 +8,20 @@ from shared.error.exceptions import (
     ServerErrorException,
     ObjectNotFound,
 )
-import abc
+from core.light_curve.infrastructure.repository import (
+    PSQLRepository,
+    MongoRepository,
+)
 
 
-class DetectionRepository(metaclass=abc.ABCMeta):
-    def get(self, object_id):
+class DetectionRepository:
+    def get(self, object_id, survey_id):
         raise NotImplementedError()
 
 
-class PSQLDetectionRepository(DetectionRepository):
+class PSQLDetectionRepository(DetectionRepository, PSQLRepository):
     def __init__(self, db: SQLConnection):
         self.db = db
-
-    def _get_object_by_id(self, object_id: str, survey_id: str):
-        try:
-            query_result = (
-                self.db.query(psql_models.Object)
-                .filter(psql_models.Object.oid == object_id)
-                .one_or_none()
-            )
-
-            if query_result:
-                return Success(query_result)
-            else:
-                return Failure(
-                    ClientErrorException(
-                        ObjectNotFound(
-                            object_id=object_id, survey_id=survey_id
-                        )
-                    )
-                )
-        except Exception as e:
-            return Failure(ServerErrorException(e))
 
     def get(self, object_id, survey_id):
         astro_obj = self._get_object_by_id(object_id, survey_id)
@@ -52,27 +33,9 @@ class PSQLDetectionRepository(DetectionRepository):
             return astro_obj
 
 
-class MongoDetectionRepository(DetectionRepository):
+class MongoDetectionRepository(DetectionRepository, MongoRepository):
     def __init__(self, db: MongoConnection):
         self.db = db
-
-    def _get_object(self, object_id, survey_id):
-        try:
-            astro_object = self.db.query().find_one(
-                model=mongo_models.Object, filter_by={"oid": object_id}
-            )
-            if astro_object:
-                return Success(astro_object)
-            else:
-                return Failure(
-                    ClientErrorException(
-                        ObjectNotFound(
-                            object_id=object_id, survey_id=survey_id
-                        )
-                    )
-                )
-        except Exception as e:
-            return Failure(ServerErrorException(e))
 
     def _get_detections(self, object_id):
         try:
